Update: Refactor Codes | Update: Change Segue | Update: Textfield interaction
